+++ CrowdControl/CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+# 0.12
+
+ - Mavenize!
+ - Default constructor, go!
+
 # 0.11
 
 Trying to get this bad boy functional! Creature spawn is in!